Revert "Add the launch browser flag to CLI" (#24)
Reverts Automattic/wordpress-playground-private#18 which adds a
platform-level feature "open in the system browser".

It is a useful feature; it just doesn't seem like a good fit at the
platform level.

There are too many OS and JS runtime combinations that we can reasonably
test. It's easy to imagine reports saying it doesn't work on Node 18 on
Linux Arch, and there's not much value in spending time solving this
type of error. Case in point: the original PR already adds an exception
for GitHub codespaces. We would inevitably see more exceptions and
special rules for other online platforms. Let's not go there at all and
leave it to the API consumers.

Every CLI consumer is free to implement that feature on their own or use
a ready package. Let's leave it to them. For the Playground platform,
it's less about "what can we add" and more about "what can we leave
out?"

cc @bgrgicak @zaerl