Add remote data blocks to the allowlist (#30)
It adds [Remote Data
Blocks](https://github.com/Automattic/remote-data-blocks) to the allowed
repositories for plugin proxy.

## Motivation for the change, related issues

In [this PR](https://github.com/Automattic/remote-data-blocks/pull/347),
we are attempting to add a live playground link for each PR so that it's
easy to test out the proposed changes. The inspiration for this came
from [the woocommerce repo](https://github.com/woocommerce/woocommerce)
that had the same feature.

## Implementation details

The code change is the same as what was done for woocommerce, and sensei
in order to support the same functionality.

## Testing Instructions (or ideally a Blueprint)

- Create a GitHub token with the "repo" scope checked for testing.
- Add the token hard coded in the
`packages/playground/website/public/plugin-proxy.php`.
- Serve the `wordpress-playground/packages/playground/website/public`
with a PHP server.
- Request

`/plugin-proxy.php?org=Automattic&repo=remote-data-blocks&workflow=Build%20Live%20Branch&artifact=remote-data-blocks-347&pr=347`
- Check that the request works properly.

Note: Depending on if you have permission to generate the right token or
not, the request will fail with a 401. I verified where the failure was
to ensure it actually tried to make the request to the correct repo
rather than rejecting it outright.