OPFS site creation: Use the same "is this directory a site?" check as the site list (#65)
Site list and site creation use different criteria for checking if an
OPFS directory represents a site.

* The site list displays OPFS directories with `wp-runtime.json` file
inside
* New site creation assumes every OPFS directory represents a site and
refuses to create a site with a given sluf if a corresponding directory
exists – even if there's no `wp-runtime.json` inside.

This means a user may not see the site in the list and still be unable
to save it under the requested slug.

This PR adjusts the check during new site creation to test for the
presence of `wp-runtime.json`.

CC @brandonpayton – what do you think?