Update Emscripten version to 4.0.5 (#93)
## Motivation for the change, related issues

[Emscripten merged a fix for
`statfs`](https://github.com/emscripten-core/emscripten/pull/23381) in
`4.0.1`, so this PR updates Emscripten to allow us to run `statfs` in
PHP functions like `disk_total_space`.

## Implementation details

This PR updates the version of Emscripten to `4.0.5`, rebuilds all
PHP-wasm libraries, and recompiles PHP-wasm.

While rebuilding PHP-wasm libraries there was a bug in the `libcurl`
build script because it expected `libopenssl` to be located in
`libopenssl/asyncify/dist`, but in reality `libopenssl` was located in
`libopenssl/asyncify/dist/1.1.0h` (similarly for JSPI).
We have two versions of `libopenssl` because it [was requires for PHP
7.0](https://github.com/wordpress/wordpress-playground/pull/2038).
This PR drops `libopenssl 1.1.0h` support because [Playground doesn't
support PHP 7.0
anymore.](https://github.com/Automattic/wordpress-playground-private/pull/74)

## Testing Instructions (or ideally a Blueprint)

- CI