Fix changelog version order (#100)
## Motivation for the change, related issues

When manually updating the older changelog versions, the version info
gets added to the top of the changelog instead of overwriting the
existing entry.

This PR doesn't address the root cause, it only updates the change log,
so versions are listed in the correct order.
I opened an issue to track the bug
https://github.com/Automattic/wordpress-playground-private/issues/101.