Remove unused dependencies (#116)
## Motivation for the change, related issues

We have a lot of dependencies which makes upgrading difficult. See #106
for more context. This PR removes the ones that seemed unused and moves
a few entries over to the devDependencies list.

After this PR, we could see which deps give us the most headache and
look into removing them in one way or another.

## Testing Instructions (or ideally a Blueprint)

Cross your fingers and wait for CI.

---------

Co-authored-by: Brandon Payton <brandon@happycode.net>
Co-authored-by: Wojtek Naruniec <wojtek@naruniec.me>
Co-authored-by: Bero <berislav.grgicak@gmail.com>