Add local package repository script (#142)
## Motivation for the change, related issues

It's hard to test unpublished packages, but sometimes it's required, so
we can test how a future Playground release will affect projects that
use our packages.

While I was working on testing an issue, I ended up building this simple
script that creates a local package repository with 0 dependencies.

@psrpinto explored testing packages using `verdaccio` in
https://github.com/Automattic/wordpress-playground-private/pull/82, and
I'm not sure which option is better.

## Implementation details

This PR adds a script for starting a local package repository.
It builds a new version of self-hosted packages using our usual process
and then it serves these files using the Playground CLI.

## Testing Instructions (or ideally a Blueprint)

- Run `npm run local-package-repository`
- Try updating a project that you have locally using the URL outputted
by the command