[Website] Renaming stored Playgrounds (#2486)
## Motivation for the change, related issues

A limited version of
https://github.com/WordPress/wordpress-playground/pull/2124 that only
enables renaming stored Playgrounds. It does not enforce choosing a name
when the Playground is first saved – it reduces the size of the PR and
makes testing and reviewing easier (the original PR is larger and breaks
the E2E tests).

## Testing Instructions (or ideally a Blueprint)

This PR ships an E2E test.

For manual testing:

- Checkout this branch
- Start the dev server npm run dev
- Go to the Playground manager
- Click save on a temporary site
- Click on the kebab menu of that site and click Rename
- Rename the site and confirm the changes are saved
- Refresh the page and confirm the new name is still there