[Website] Apply base URLs to relative redirection URLs (#2595)
Fixes relative redirects support in wp-admin.

Technically, this PR ensures that redirection URLs handled in
`convertFetchEventToPHPRequest` are combined with a base URL before
passing them on to the browser.

Fixes #2578

## Testing Instructions (or ideally a Blueprint)

Go to wp-admin and create a new user. Confirm you are redirected back to
wp-admin in the end and not to a 404 page