[Xdebug Bridge] Fix issue if non existent file is read (#2625)
## Motivation for the change, related issues

Based on
https://github.com/WordPress/wordpress-playground/issues/2601#issuecomment-3276923541

When `readPHPFile` tries to read a non-existent file based on a wrong
path, it crashes with code `44` and message :

```
Error [ErrnoError]: Could not read "": There is no such file or directory OR the parent directory does not exist.
```

This pull request aims to bypass those non-existent files and focus on
the existing ones. If the file reading attempt crashes during
`sendScriptToCDP` a warning is logged in the Devtools console with the
error message while the process is still running.

## Implementation details

- Addition of a `try..catch` statements around `this.readPHPFile` and
what follows.
- Implementation of a new `Log.entryAdded` with the `ErrnoError` message
in the `catch` part.
- Implementation of a test.

## Testing Instructions (or ideally a Blueprint)

CI