[CLI] Respond with HTTP 500 response when the request handler throws an error (#2715)
## Motivation for the change, related issues

Without this PR, an exception thrown by `PHP.run()` would cause the CLI
thread to exit with error code 1 due to an UnhandledPromiseRejection
error. With this PR, the exception will be logged and the user will see
a basic page communicating HTTP error 500 (internal server error).

## Testing Instructions (or ideally a Blueprint)

This is challenging. I've added a test, hoping it would pass, but
something in the setup makes it fail (because PHP.run() refuses to throw
an exception). I wonder how could we add a unit test for this
@brandonpayton

---------

Co-authored-by: Brandon Payton <brandon@happycode.net>