[Website] Make deployment script more usable for self-hosting (#2764)
## Motivation for the change, related issues

The Telex project wants to self-host Playground and pin it to a specific
version. This is a PR to make deployment to WP Cloud easier for folks
like them.

## Implementation details

This PR:
- Adds a `GIT_REF_TO_DEPLOY` var so users can configure the workflow
environment to deploy a specific branch, tag, or revision.
- Adds a `CORS_PROXY_URL` var to allow folks to choose their CORS proxy.
- Renames the workflow file from build-website.yml to deploy-website.yml
to better match its purpose.
- Makes names within the workflow more generic with self-hosting in
mind. For example, "Deploy to playground.wordpress.net" becomes "Deploy
Playground website".
- Updates to actions/checkout@v5 since we might as well make the updates
when touching this workflow.
- Deletes the private deploy workflow because it is redundant and no
longer used.

## Testing Instructions (or ideally a Blueprint)

- Test deploying to other WP Cloud site in Playground fork. This is
already done
[here](https://github.com/brandonpayton/wordpress-playground/actions/runs/18394052721).
- Test deploying to playground.wordpress.net after merge.