[Blueprints] ensure git:directory resource returns non-empty-name (#2779)
## Motivation for the change, related issues

Before this PR, this step would fail:

```js
{
		"step": "installPlugin",
		"pluginData": {
			"resource": "git:directory",
			"url": "https://github.com/juanma-wp/meetup-cheltenham-workshop/",
			"ref": "final",
			"path": "/"
		},
		"options": {
			"activate": true
		}
},
```

This is because `installPlugin` names the plugin directory inside
`wp-content` after the filename of the resolved resource, and the
`git:directory` resource would use `path` to generate a filename called
`/`.

With this PR, `git:directory` generates a directory name using the full
set of information from the resource spec: URL, ref, path. The names are
now more like this:
`https-github.com-WordPress-wordpress-playground-trunk`.

## Testing Instructions (or ideally a Blueprint)

CI – this PR comes with tests

cc @juanmaguitar