[Website] Remove client package from the list of offline mode assets (#2841)
## Motivation for the change, related issues

When loading Playground I noticed an _Uncaught (in promise) TypeError:
Failed to execute 'addAll' on 'Cache': Request failed_ error caused by
offline mode caching.
The error was triggered because _blueprints-DY7omvFh.cjs_ and
_blueprints-Dg6flSLH.js_ are located in the `/client/` subdirectory on
Playground.WordPress.net while the [list of offline mode
assets](https://playground.wordpress.net/assets-required-for-offline-mode.json)
expects these files to be in the root folder.

To address the issue, this PR removes the _client_ package from the list
of offline assets.
We are also removing the _php-playground.html_ page from the list
because it depends on the client package but isn't part of the web
version of Playground.

## Testing Instructions (or ideally a Blueprint)

- Follow the [Test offline support
instructions](https://github.com/WordPress/wordpress-playground/blob/trunk/README.md#test-offline-support)
and ensure offline mode is working