[CLI] Fix temp dir cleanup when Playground CLI server is killed (#2836)
## Motivation for the change, related issues

Today, when Playground CLI is killed via Ctrl-C (SIGINT), the temp dir
is left behind. This is a problem because Playground CLI server runs
until it is killed in such a manner.

## Implementation details

This PR adds explicit temp dir cleanup when receiving the signals SIGINT
or SIGTERM.

## Testing Instructions (or ideally a Blueprint)

- Run Playground CLI server in the console with the `--verbose=debug`
flag.
- Note the temp dir it created and printed in the CLI output
- Kill the server with Ctrl-C
- Confirm that the director no longer exists

Prior to this patch, the temp dir still exists after this test. After
this patch, the temp dir is removed.