[CLI] Fix null and "latest" WP version resolution and improve unzip error message (#2889)
## Motivation for the change, related issues

Today, while there is an active WP 6.9 release candidate, Playground CLI
is using the release candidate as "latest". As a user, I would expect
"latest" to remain the latest production release. This PR makes that
change.

In addition, when passing `null` as a CLI `wp` argument, Playground CLI
attempts to download `https://wordpress.org/wordpress-null.zip`. In that
case, WordPress.org outputs a "Release not found" generic text file that
is cached as if it is a good release. This PR addresses that case as
well.

## Implementation details

This PR adjusts the version selection logic to skip release candidates
and treat the latest production release as "latest". It also updates the
tests to include this scenario.

This PR addresses the null version issue by:

1. Resolving `null` versions to `latest`
2. Output additional information in the unzip function

## Testing Instructions (or ideally a Blueprint)

- CI

---------

Co-authored-by: Francesco Bigiarini <francesco.bigiarini@gmail.com>