Add file locking integration tests for php-wasm/node (#2908)
## Motivation for the change, related issues

We have automated tests for our file lock manager but no integration
tests to demonstrate locking is working between `@php-wasm/node`
instances. Let's fix that.

## Implementation details

This PR:
- Adds automated tests for file locking within `@php-wasm/node`
- Fixes a bug discovered by the additional tests: releasing flock()
locks when the associated file descriptor is closed.

The tests coordinate between two or more concurrent `@php-wasm`
instances using a simple text file to gate the stages. For example, for
php instances php1 and php2 testing exclusive locking:
- File initially contains "php1-locking"
- php1 starts without waiting
- php2 waits until file contains "php1-waiting-for-php2-to-try"
- php1 obtains an exclusive lock on a SQLite DB and sets file contents
to "php1-waiting-for-php2-to-try"
- php1 waits until file contains "php2-ready-for-unlock"
- php2 attempts to write to DB, records the outcome, and sets file
contents to "php2-ready-for-unlock"
- php2 waits until file contains "php1-unlocked"
- php1 unlocks the DB, sets file contents to "php1-unlocked", and exits
- php2 repeats its attempt to right the DB, records the outcome, echoes
its results, and exits
- the test case asserts expected outcome

Maybe there is a better way to coordinate between instances, but so far,
this seemed the most straightforward.

Once these tests are fleshed out, we can probably add nearly identical
tests for Playground CLI, but run the scripts via HTTP requests.

## Testing Instructions (or ideally a Blueprint)

- CI