Add `SQLITE_ENABLE_COLUMN_METADATA` and update SQLite (#2948)
## Motivation for the change, related issues
Our SQLite builds don't set the `SQLITE_ENABLE_COLUMN_METADATA` flag,
which results in some metadata missing when
[`PDOStatement::getColumnMeta`](https://www.php.net/manual/en/pdostatement.getcolumnmeta.php)
is used. This results in issues like missing row editing controls in
phpMyAdmin and possibly other problems.

This flag seems to be set in most libsqlite3 builds:
https://www.rogerbinns.com/blog/sqlite-configuration.html

Therefore, I think it's safe to add it and, from the SQLite side, assume
that it's always used.

## Implementation details
I added the flag, updated SQLite, rebuilt it, and rebuilt all PHP
builds.

## Testing Instructions (or ideally a Blueprint)
I added tests for the `SQLITE_ENABLE_COLUMN_METADATA` and they need to
pass.