[CI] Enable trusted publisher-based publishing (#3048)
## Motivation for the change, related issues

Rewires the publish package's job to use trusted publisher's setup. Here
is why it wasn't used before:

* Injecting `NPM_TOKEN` and writing `.npmrc` disables OIDC and forces
legacy auth.
* Checkout using `github.event.pull_request.head.ref` – releases need to
be deterministic and always run from trunk. This change limits our
ability to publish a preview version of the packages with a different
dist tag, which seems fine in practice since we haven't used it at all.

Technically unrelated, but I've coupled this change with this PR: we're
no longer rewriting the Git remote with a secret token. The baked in
GITHUB_TOKEN already has the required permissions.