[PHP] Remove unnecessary createRequire() from @php-wasm/web build scripts (#3076)
## Motivation for the change, related issues

Attempts to solve the `createRequire is not e function` issue reported
at https://github.com/WordPress/wordpress-playground/issues/3074

## Implementation details

Removes a unused `createRequire` definition from the ESM build script
for every PHP web version

## Testing Instructions (or ideally a Blueprint)

None, we'll need to publish the packages and confirm they work. It seems
terrible, but they're already broken at the moment so we're not risking
anything. As a follow-up, we'll need to find a way to test the built web
packages before publishing them.